pyproject.toml
[project]
name = "hil-bench"
version = "0.1.0"
description = "Public HIL-Bench runtime package"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "datasets>=2.20.0",
    "litellm>=1.81.11",
    "sweagent",
    "swebench>=4.1.0",
    "flask>=3.0.0",
    "chromadb==1.3.3",
    "sentence_transformers>=5.3.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
]

[project.scripts]
hil = "hil_bench.cli:main"

[tool.uv]
package = true

[tool.uv.workspace]
members = ["SWE-agent"]

[tool.uv.sources]
sweagent = { workspace = true }

[tool.setuptools.packages.find]
include = ["hil_bench*"]
